% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDstat.R
\name{DDstat}
\alias{DDstat}
\title{The Dutta & Datta statistic}
\usage{
DDstat(mdat, iICGs = TRUE, get_var = TRUE)
}
\arguments{
\item{mdat}{a numeric matrix with columns names 'cID', 'Y', and 'G' that denotes the cluster ID, response, and group indicator, respectively.}

\item{iICGs}{a logical argument indicating if the data has incomplete ICG structure. Use the \code{crspack::chkICG} function, as needed.}

\item{get_var}{a logical argument indicating whether the variance of the estimate should be obtained. Default is \code{TRUE}.}
}
\value{
a list that comprise the Dutta & Datta test statistic, expected value, variance (optional), and Z-score (optional).
}
\description{
Computes the Dutta & Datta statistic.
}
\details{
Estimates the test statistic, expected value and variance of \insertCite{dutta2016;textual}{crspack}.
}
\examples{
## simple data simulation
m <- 3
mdat <- lapply(1:m, function(x){
  ni <- rpois(n=1, lambda=4)+2 # cluster size
  cID <- rep(x=x, times=ni) # cluster ids
  nu <- rnorm(n=1, mean=0, sd=0.25) # cluster-level random effect
  G <- rbinom(n=ni, size=1, prob=0.5) # grouping variable
  err <- rnorm(n=ni, mean=0, sd=0.25)
  err[which(G == 1)] <- rnorm(n=sum(G == 1), mean=0.15, sd=0.25)
  Y <- 0.5 + nu + err # response
  data.frame(cID=cID, Y=Y, G=G)
})
mdat <- do.call(rbind, mdat)
mdat <- apply(as.matrix(mdat),2,as.numeric)

## Estimate the DD statistic
DDstat(mdat=mdat, iICGs=TRUE, get_var=TRUE)
}
\references{
{
\insertRef{dutta2016}{crspack}
}
}
\author{
Samuel Anyaso-Samuel, Somnath Datta
}
