% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkICG.R
\name{chkICG}
\alias{chkICG}
\title{Incomplete ICG structure}
\usage{
chkICG(mdat)
}
\arguments{
\item{mdat}{a numeric matrix with columns names 'cID', 'Y', and 'G' that denotes the cluster ID, response, and group indicator, respectively.}
}
\value{
a boolean, indicating if data has incomplete ICG structure.
}
\description{
Incomplete ICG structure
}
\details{
Checks if data has incomplete ICG structure.
}
\examples{
## simple data simulation
m <- 3
mdat <- lapply(1:m, function(x){
  ni <- rpois(n=1, lambda=4)+2 # cluster size
  cID <- rep(x=x, times=ni) # cluster ids
  nu <- rnorm(n=1, mean=0, sd=0.25) # cluster-level random effect
  G <- rbinom(n=ni, size=1, prob=0.5) # grouping variable
  err <- rnorm(n=ni, mean=0, sd=0.25)
  err[which(G == 1)] <- rnorm(n=sum(G == 1), mean=0.15, sd=0.25)
  Y <- 0.5 + nu + err # response
  data.frame(cID=cID, Y=Y, G=G)
})
mdat <- do.call(rbind, mdat)
mdat <- apply(as.matrix(mdat),2,as.numeric)

## Checks if data has incomplete ICG structure
chkICG(mdat=mdat)
}
\references{
{
\insertRef{dutta2016}{crspack}
}
}
\author{
Samuel Anyaso-Samuel, Somnath Datta
}
